requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
